app/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.8
    # via requests
click==8.0.3
    # via pip-tools
django==3.2.4
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
packaging==21.3
    # via pytest
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
psycopg2-binary==2.9.1
    # via -r requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.6
    # via packaging
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.4.0
    # via -r requirements.in
pytz==2021.3
    # via django
requests==2.26.0
    # via -r requirements.in
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pytest
tomli==1.2.2
    # via pep517
urllib3==1.26.7
    # via requests
wheel==0.37.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
